Move all run-pass synchronization primitive tests to sync.rs.
