--- tests/run-pass/concurrency/locks.stderr
@@ -1,2 +1,0 @@
-warning: thread support is experimental. For example, Miri does not detect data races yet.
-
