+++ tests/run-pass/concurrency/sync.stderr
