+++ tests/run-pass/concurrency/sync.stdout
